refactor: migrate client components to separate files for better code organization
